Split cookie auth helpers into one export per file
