+++ packages/twenty-front/src/modules/apollo/services/apollo.factory.ts
@@ -20,11 +20,9 @@ import { retryWithBackoff } from '~/utils/retryWithBackoff';
 
 import { REST_API_BASE_URL } from '@/apollo/constant/rest-api-base-url';
 import { type ApolloManager } from '@/apollo/types/apolloManager.interface';
-import {
-  getIsCookieAuthActive,
-  setIsCookieAuthActive,
-} from '@/apollo/utils/cookieAuthActive';
+import { getIsCookieAuthActive } from '@/apollo/utils/getIsCookieAuthActive';
 import { getTokenPair } from '@/apollo/utils/getTokenPair';
+import { setIsCookieAuthActive } from '@/apollo/utils/setIsCookieAuthActive';
 import { isUnauthenticatedGraphQLError } from '@/apollo/utils/isUnauthenticatedGraphQLError';
 import { loggerLink } from '@/apollo/utils/loggerLink';
 import { StreamingRestLink } from '@/apollo/utils/streamingRestLink';
